Split maintext from mainwindow
The single MainText widget will need to be accessed by many
features, e.g. Search. However, the search feature should not
need to import the whole of mainwindow, which includes menus,
status bar, etc.
